vector-search/requirements.txt
fastapi
uvicorn
faiss-cpu
requests
pydantic 